changing ints to bigint because CRDB uses 64 bit to mean int (#103)

+++ internal/migrator/types.go
@@ -158,7 +158,7 @@ func (c Column) ConvertPostgresDataTypeToSqlserver(columnName string) string {
 			convertedDataType.WriteString("varchar(max)")
 		}
 	case "NUMBER":
-		convertedDataType.WriteString("int")
+		convertedDataType.WriteString("bigint")
 	case "BOOLEAN":
 		convertedDataType.WriteString("bit")
 	case "TIMESTAMP_TZ":
